Create README.md
Includes various small c++ projects. All can be run using opening the executable in the respective bin folders
---------------------------------------------------------------------------------------------------------------------------------------------
Number maze game and solver
A c++ implementation of the number maze game: http://logicmazes.com/n1mz.html.
The game reads from a text file in which the first row consists of 6 numbers 
     -the first two numbers indicative of a matrix size n by m.
     -the second two numbers indicative of the starting point by row, col, usually position 0, 0 however can be any position with n by m
     -the third two numbers indicative of the ending point by row, col likewise.
The remaining rows and columns define the matrix of which the puzzle will represent.
The object of the game is to reach the end goal, the position defined by the third two numbers from the first row.
The player may move only the amount of spaces defined by the number the player is currently on.
The player moves with the keys wasd. 
There is a solver included which finds all possible solution paths.
---------------------------------------------------------------------------------------------------------------------------------------------
8 puzzle game
---------------------------------------------------------------------------------------------------------------------------------------------
Hopfield network test
---------------------------------------------------------------------------------------------------------------------------------------------
A small scene in opengl, randomly generated trees.
Pressing the arrow keys changes view, pressing r randomizes tree generation